// frontend/src/app/administracion-oee/planes-tic/models/planes-tic.model.ts
/* import { Oee } from "../../oee/models/oee.model"; */

export class PlanesTic {
    idPlan: number;
    anho: string;
    fecha: string;
    linkPlan: string;
    cantidadFuncionariosTic: string;
    cantidadFuncionariosAdmin: string;
    presupuestoTicAnual: string;
    estado: boolean;
    /* oee: Oee[];
    estado: string;
    anhoCreacion: number;
    poseeCodigoFuente: boolean; */
}